TASK: Rename to `ResetupAndReplayContentGraphUpgrade`
